Refactor BaseDataSource.isConnectable().
