+++ dolphinscheduler-dao/src/main/java/org/apache/dolphinscheduler/dao/datasource/OracleDataSource.java
@@ -21,10 +21,6 @@ import org.apache.dolphinscheduler.common.utils.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-
 /**
  * data source of Oracle
  */
@@ -33,7 +29,7 @@ public class OracleDataSource extends BaseDataSource {
 
     /**
      * gets the JDBC url for the data source connection
-     * @return
+     * @return jdbc url
      */
     @Override
     public String getJdbcUrl() {
@@ -52,24 +48,10 @@ public class OracleDataSource extends BaseDataSource {
     }
 
     /**
-     * test whether the data source can be connected successfully
-     * @throws Exception
+     * @return driver class
      */
     @Override
-    public void isConnectable() throws Exception {
-        Connection con = null;
-        try {
-            Class.forName(Constants.COM_ORACLE_JDBC_DRIVER);
-            con = DriverManager.getConnection(getJdbcUrl(), getUser(), getPassword());
-        } finally {
-            if (con != null) {
-                try {
-                    con.close();
-                } catch (SQLException e) {
-                    logger.error("Oracle datasource try conn close conn error", e);
-                }
-            }
-        }
-
+    public String driverClassSelector() {
+        return Constants.COM_ORACLE_JDBC_DRIVER;
     }
 }
